doFillCredentialsItems and  doCheckCredentials
Adding doFillCredentialsItems and  doCheckCredentials to Asoc Analyzer.
Removing the same from the AppScanBuildStep
