+++ src/main/java/com/hcl/appscan/jenkins/plugin/builders/AppScanBuildStep.java
@@ -384,42 +384,6 @@ public class AppScanBuildStep extends Builder implements SimpleBuildStep, Serial
         public String getDisplayName() {
             	return Messages.label_build_step();
         }
-    	
-    	public ListBoxModel doFillCredentialsItems(@QueryParameter String credentials, @AncestorInPath ItemGroup<?> context) {
-    		//We could just use listCredentials() to get the ListBoxModel, but need to work around JENKINS-12802.
-    		ListBoxModel model = new ListBoxModel();
-    		List<ASoCCredentials> credentialsList = CredentialsProvider.lookupCredentials(ASoCCredentials.class, context,
-    				ACL.SYSTEM, Collections.<DomainRequirement>emptyList());
-    		boolean hasSelected = false;
-    		
-    		for(ASoCCredentials creds : credentialsList) {
-    			if(creds.getId().equals(credentials))
-    				hasSelected = true;
-    			String displayName = creds.getDescription();
-    			displayName = displayName == null || displayName.equals("") ? creds.getUsername() + "/******" : displayName; //$NON-NLS-1$
-    			model.add(new ListBoxModel.Option(displayName, creds.getId(), creds.getId().equals(credentials))); //$NON-NLS-1$
-    		}
-    		if(!hasSelected)
-    			model.add(new ListBoxModel.Option("", "", true)); //$NON-NLS-1$ //$NON-NLS-2$
-    		return model;
-    	}
-    	
-    	
-    	
-    	
-    	
-    	public FormValidation doCheckCredentials(@QueryParameter String credentials, @AncestorInPath ItemGroup<?> context) {
-    		if(credentials.trim().equals("")) //$NON-NLS-1$
-    			return FormValidation.errorWithMarkup(Messages.error_no_creds("/credentials")); //$NON-NLS-1$
-    		
-    		IAuthenticationProvider authProvider = new JenkinsAuthenticationProvider(credentials, context);
-    		if(authProvider.isTokenExpired())
-    			return FormValidation.errorWithMarkup(Messages.error_token_expired("/credentials")); //$NON-NLS-1$
-    			
-    		return FormValidation.ok();
-    	}
-    	
-    	
     }
 }
