+++ src/main/java/com/hcl/appscan/jenkins/plugin/scanners/StaticAnalyzer.java
@@ -6,7 +6,13 @@
 
 package com.hcl.appscan.jenkins.plugin.scanners;
 
+import com.cloudbees.plugins.credentials.CredentialsProvider;
+import com.cloudbees.plugins.credentials.domains.DomainRequirement;
+import com.hcl.appscan.jenkins.plugin.Messages;
+import com.hcl.appscan.jenkins.plugin.auth.ASoCCredentials;
+import com.hcl.appscan.jenkins.plugin.auth.JenkinsAuthenticationProvider;
 import com.hcl.appscan.sdk.CoreConstants;
+import com.hcl.appscan.sdk.auth.IAuthenticationProvider;
 import com.hcl.appscan.sdk.scanners.sast.SASTConstants;
 import java.util.HashMap;
 import java.util.Map;
@@ -16,8 +22,14 @@ import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.QueryParameter;
 
 import hudson.Extension;
+import hudson.model.ItemGroup;
+import hudson.security.ACL;
 import hudson.util.FormValidation;
+import hudson.util.ListBoxModel;
 import hudson.util.VariableResolver;
+import java.util.Collections;
+import java.util.List;
+import org.kohsuke.stapler.AncestorInPath;
 import org.kohsuke.stapler.DataBoundSetter;
 
 public class StaticAnalyzer extends Scanner {
@@ -76,5 +88,35 @@ public class StaticAnalyzer extends Scanner {
 	    	public FormValidation doCheckTarget(@QueryParameter String target) {
 	    		return FormValidation.validateRequired(target);
 	    	}
+		
+		public ListBoxModel doFillCredentialsItems(@QueryParameter String credentials, @AncestorInPath ItemGroup<?> context) {
+            //We could just use listCredentials() to get the ListBoxModel, but need to work around JENKINS-12802.
+            ListBoxModel model = new ListBoxModel();
+            List<ASoCCredentials> credentialsList = CredentialsProvider.lookupCredentials(ASoCCredentials.class, context,
+                            ACL.SYSTEM, Collections.<DomainRequirement>emptyList());
+            boolean hasSelected = false;
+
+            for(ASoCCredentials creds : credentialsList) {
+                    if(creds.getId().equals(credentials))
+                            hasSelected = true;
+                    String displayName = creds.getDescription();
+                    displayName = displayName == null || displayName.equals("") ? creds.getUsername() + "/******" : displayName; //$NON-NLS-1$
+                    model.add(new ListBoxModel.Option(displayName, creds.getId(), creds.getId().equals(credentials))); //$NON-NLS-1$
+            }
+            if(!hasSelected)
+                    model.add(new ListBoxModel.Option("", "", true)); //$NON-NLS-1$ //$NON-NLS-2$
+            return model;
+        }
+                
+        public FormValidation doCheckCredentials(@QueryParameter String credentials, @AncestorInPath ItemGroup<?> context) {
+            if(credentials.trim().equals("")) //$NON-NLS-1$
+                    return FormValidation.errorWithMarkup(Messages.error_no_creds("/credentials")); //$NON-NLS-1$
+
+            IAuthenticationProvider authProvider = new JenkinsAuthenticationProvider(credentials, context);
+            if(authProvider.isTokenExpired())
+                    return FormValidation.errorWithMarkup(Messages.error_token_expired("/credentials")); //$NON-NLS-1$
+
+            return FormValidation.ok();
+        }
 	}
 }
